모바일개발을 위한 디렉터리 구조 변경
