--- story-front/components/Footer/Mobile/index.tsx
@@ -1,8 +1,0 @@
-const Footer: React.FC = () => {
-    return (
-        <div>
-        </div>
-    )
-}
-
-export default Footer
