+++ story-front/components/Footer/css/index_mobile.module.css
@@ -1,0 +1,64 @@
+.footer{
+    margin-top:0px;
+    position:relative;
+    width:100%;
+    height:170px;
+    border:#ddd solid;
+    border-width:1px;
+    border-left:0px;
+    border-right:0px;
+    border-bottom:0px;
+    font-size:0.8em;
+    color:#555;
+}
+.footer_container{
+    margin:0 auto;
+    margin-top:11px;
+    width:1124px;
+    height:150px;
+}
+.footer_inner_container_left{
+    width:674px;
+    height:100%;
+    float:left;
+}
+.footer_inner_container_right{
+     width:450px;
+    height:100%;
+    float:left;
+    text-align:center;
+}
+.footer_menu_1{
+    margin-top:5px;
+}
+
+.footer_menu_1 > a {
+    color:black;
+}
+
+.footer_menu_1 > a:hover {
+    color: var(--color-main);
+    text-decoration: none;
+}
+
+.footer_business_info_first{
+    margin-top:15px;
+    font-size:0.9em;
+}
+.footer_business_info{
+    margin-top:5px;
+    font-size:0.9em;
+}
+.footer_menu_4{
+    margin-top:20px;
+    color:#888;
+    font-size:0.9em;
+}
+.call_text{
+    font-size:2.1em;
+    font-weight:1000;
+    color:gray;
+}
+.footer_logo{
+    margin-top:5px;
+}
