+++ story-front/components/Header/index.tsx
@@ -1,2 +1,1 @@
 export { default as Header } from './Pc'
-export { default as HeaderMobile } from './Mobile'
