+++ story-front/components/Image/index.module.css
