+++ story-front/components/Product/Contents.tsx
@@ -5,9 +5,9 @@ import Summary from './Summary';
 import Content from './Content';
 import CircularProgress from '@material-ui/core/CircularProgress';
 import { ImagePopup } from '@/components/index';
-import { deepArrayCopy } from '../../util';
-import { RequestMoreProductListAction, RequestProductListAction } from '../../store/action/sagaAction';
-import { ProductListData } from '../../store/action/reducerAction';
+import { deepArrayCopy } from '@/util/index';
+import { RequestMoreProductListAction, RequestProductListAction } from '@/store/action/sagaAction';
+import { ProductListData } from '@/store/action/reducerAction';
 
 
 const ProgressBox = styled.div`
