+++ story-front/components/index.tsx
@@ -1,8 +1,8 @@
-export { Footer } from './Footer';
+export { default as Footer } from './Footer';
 export { Header } from './Header';
 export { default as Cardslider } from './Slider/Cardslider';
-export { Image } from './Image';
-export { Label } from './Label';
+export { default as Image } from './Image';
+export { default as Label } from './Label';
 export { default as PContents } from './Product/Contents';
 export { default as PFilter } from './Product/Filter';
 export { default as PulseBox } from './PulseBox';
